requirements.txt
dlib==19.24.0
Flask==2.2.2
numpy==1.23.5
opencv_python==4.7.0.72
Pillow==9.5.0
